backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
elevenlabs==1.50.7
python-dotenv==1.0.1
pydantic==2.10.6
